Split utils test too
